chore: auth_config file rename, multi_prisma_client generation
