+++ apps/auth/src/auth/data/repo/user.repository.ts
@@ -1,9 +1,10 @@
 import { Inject, Injectable, OnModuleDestroy, OnModuleInit } from "@nestjs/common";
-import { PrismaClient } from "@prisma/client";
+
 import { User } from "../../model/user.model";
 import { IDatasource } from "@app/common/datasource_helper/datasource.interface";
 import { QueryHelper } from "@app/common/datasource_helper/query_helper";
 import { RequestValidationException } from "@app/common/errors/request_validation_exception";
+import { PrismaClient } from "apps/auth/prisma/generated/prisma_auth_client";
 
 export abstract class IUserRepository {
     abstract createUser(userInfo: User): Promise<User>
