+++ apps/auth/src/auth/service/guard/jwt.service.ts
@@ -3,7 +3,7 @@ import { ConfigService } from "@nestjs/config";
 import { PassportStrategy } from "@nestjs/passport";
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { JwtPayload } from "../../model/jwt_payload.model";
-import { ACCESS_TOKEN } from "apps/auth/configuration";
+import { ACCESS_TOKEN } from "apps/auth/auth_configuration";
 import { IUserRepository, UserRepository } from "../../data/repo/user.repository";
 import { RequestValidationException } from "@app/common/errors/request_validation_exception";
 
@@ -28,4 +28,4 @@ export class AccessTokenStretegy extends PassportStrategy(Strategy, "jwt") {
         throw new RequestValidationException({ message: "Unauthorized error", statusCode: 401 })
     }
 
-} 
+}
